requirements.txt
azure-ai-formrecognizer==3.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
azure-search-documents==11.4.0 